error handling 세분화
