+++ ridi_oauth2/introspector/exceptions.py
@@ -10,5 +10,25 @@ class InvalidToken(Exception):
     pass
 
 
-class InvalidPublicKey(Exception):
+class PublicKeyException(Exception):
+    pass
+
+
+class InvalidPublicKey(PublicKeyException):
+    pass
+
+
+class NotExistedKey(PublicKeyException):
+    pass
+
+
+class PublicRequestException(PublicKeyException):
+    pass
+
+
+class AccountServerException(PublicRequestException):
+    pass
+
+
+class ClientRequestException(PublicRequestException):
     pass
